truncations to common mod
